central_server/flask_app/requirements.txt
APScheduler==3.11.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
influxdb-client==1.48.0
numpy==2.2.4
pandas==2.2.3
paramiko==3.5.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
requests==2.32.3
